+++ docs/source/changelog.rst
@@ -18,6 +18,15 @@ Fixed
 .. _#593: https://github.com/Bogdanp/dramatiq/pull/593
 .. _@pahrohfit: https://github.com/pahrohfit
 
+Added
+^^^^^
+
+* CLI watcher now supports setting include and exclude patterns (`#594`_,
+  `@nhairs`_)
+
+.. _#594: https://github.com/Bogdanp/dramatiq/issues/594
+.. _@nhairs: https://github.com/nhairs
+
 `1.15.0`_ -- 2023-10-23
 -----------------------